Move PrettyPrint to its own file
